backend/requirements.txt
flask==2.1.0
werkzeug==2.1.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
python-dotenv==0.20.0
pytz==2023.3
requests==2.26.0
email-validator==1.1.3
gunicorn==20.1.0